benchmarking/requirements.txt
pycapnp==1.3.0
raylib-py==5.0.0
